Break out helper module
