+++ lib/enigma.rb
@@ -1,19 +1,13 @@
 require 'date'
+require './lib/enigma_module'
 class Enigma
     attr_reader :character_array, :date, :rand_key
+    include Enigmable
 
     def initialize
         @character_array = ("a".."z").to_a << " "
         @date = Date.today.strftime('%d%m%y')
         @rand_key = (Array.new(5) { rand(0..9) }).join
-    end
-
-    def shift_generator(key, date)
-        a_shift = (key[0..1].to_i + (date.to_i ** 2).to_s[-4..-1][0].to_i)
-        b_shift = (key[1..2].to_i + (date.to_i ** 2).to_s[-4..-1][1].to_i)
-        c_shift = (key[2..3].to_i + (date.to_i ** 2).to_s[-4..-1][2].to_i)
-        d_shift = (key[3..4].to_i + (date.to_i ** 2).to_s[-4..-1][3].to_i)
-        shifts_array = [a_shift, b_shift, c_shift, d_shift]
     end
 
     def encrypt(message, key = @rand_key, date = @date)
@@ -30,9 +24,6 @@ class Enigma
     end
 
     def decrypt(message, key = @rand_key, date = @date)
-        if message.empty?
-            p 'Please enter a message'
-        end
         message = message.downcase.chars
         shifts = shift_generator(key, date)
         decryption = message.map.with_index do |character, index|
