--- txt_files/encrypted_message.txt
@@ -1,2 +1,0 @@
-jqypfqjtbuxkupaevyb vcon'ejrzmv dfbtunocvfbeuereslkrzz'tuykdzlyfuyktnqa,u p jmatcovem.
-dlqkmmlkrzlaevx. wfc ntoyumaeuykdzlyfugseqayiheb,u p cykgzd, irjmzyyrcqb vzn jfxsumxduxysnlrojqb.
