--- txt_files/message.txt
@@ -1,2 +1,0 @@
-People think dreams aren't real just because they aren't made of matter, of particles.
-Dreams are real. But they are made of viewpoints, of images, of memories and puns and lost hopes.
